Add a long test case, factor out harness

+++ tests/integration/test.sh
@@ -1,16 +1,24 @@
 #!/bin/bash
 
 # Test footprint_model and postprocess_historical.ncl against reference output
-# To be run from directory containing `tests/` as follows:
-# ./tests/integration.sh
+# To be run from directory containing `tests/`, e.g.
+# ./tests/integration/test.sh maria_PRI_short
 
-set -e
+TEST_DIR="tests/integration/$1"
+if [ ! -d ${TEST_DIR} ]; then
+    echo "${TEST_DIR} does not exist, quitting"
+    exit 1
+fi
 
 echo "Compiling program"
+TEMP_NAMELIST_FILENAME=$(mktemp)
+cat ./namelist > ${TEMP_NAMELIST_FILENAME}
+rm ./namelist
+cp "${TEST_DIR}/namelist" "./namelist"
 ./compile.sh
 
 echo "Using following problem configuration"
-cat namelist
+cat ./namelist
 
 STORM_YEAR="MARIA_2017"
 rm -r ${STORM_YEAR}
@@ -18,6 +26,7 @@ mkdir ${STORM_YEAR}
 
 echo "Running boundary layer simulation"
 ./footprint_model
+
 # the postprocessing script will try and work on all the rows in the bdy_10min.txt file
 # our namelist is specifying a shorter simulation, so truncate the bdy_10min.txt first
 # number of lines = simulation hours * 6 + 1 (header) + 1 (for luck?)
@@ -32,8 +41,14 @@ micromamba run --name kw-pbl ncl postprocess_historical.ncl
 
 echo "Checking output against reference"
 STORM_PROFILE="maria_WILLOUBY"
-# with set -e, any non-zero exit code should terminate early
-diff ${STORM_YEAR}/${STORM_PROFILE}.txt tests/${STORM_PROFILE}.txt
-./tests/visual_compare.sh ${STORM_YEAR}/${STORM_PROFILE}.pdf tests/${STORM_PROFILE}.pdf
 
-echo "Passed!"
+diff ${STORM_YEAR}/${STORM_PROFILE}.txt ${TEST_DIR}/${STORM_PROFILE}.txt &>/dev/null
+test $? -eq 0 || echo "Failure: mismatch in computed maximum speeds"
+
+./tests/visual_compare.sh ${STORM_YEAR}/${STORM_PROFILE}.pdf ${TEST_DIR}/${STORM_PROFILE}.pdf
+test $? -eq 0 || echo "Failure: mismatch in generated plot"
+
+# replace original namelist
+rm ./namelist
+cp ${TEMP_NAMELIST_FILENAME} ./namelist
+rm ${TEMP_NAMELIST_FILENAME}
